// src/app/models/payment.model.ts
export class Payment {
  lastDigitNumber: string;
  cardid: string;
  tokenID: string;
  addressZip: string;
  brand: string;
  expiryMonth:number;
  expiryYear:number;
  amount: number;
  userEmail: string;
  compaignID: number;
  clientID: number;
  callCenterID: number;
  senderType: string;
  recieverType: string;
  expiryDate: string;
  value: number;
  cvc: string;
  cardNumber: string;
}